Fixing my idiocy
Dropped the files in the wrong place. Eventually I'll stop being an
idiot.
